+++ code/system.ui/react.common/src/ui.dev/DevTools.Button/index.ts
@@ -1,1 +1,1 @@
-export * from './Button';
+export * from './button';
